Add friend request, accept and decline buttons
